refactor: random testutils file
